Update role dependencies
Add the cisagov/ansible-role-docker role as a dependency. Although
Docker is not directly invoked in this role the cisagov/ncats-webui
repository is a Docker configuration and using it will require that
Docker is installed and available. Adding cisagov/ansible-role-docker
as a dependency has necessitated switching to SystemD-enabled images
for the molecule configuration.